refactor plugin/macro to share majority of the code
